Add a "Reload all" action
Adds a menuitem to the file menu, allows the creation of keyboard
shortcuts, but does not add a toolbar action (no suitable icon available)
nor add a menu item to the sidebar menu.
The first loop looks for changed files and asks the user *once* if he want to
reload *all* files, the second one handles the actual reloading.